+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_meedu_videoplayer/meedu_player.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/basic_example_page.dart';
@@ -6,35 +5,41 @@ import 'package:flutter_meedu_videoplayer_example/pages/change_quality_example_p
 import 'package:flutter_meedu_videoplayer_example/pages/custom_icons_example.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/disabled_buttons_example_page.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/fullscreen_example_page.dart';
+import 'package:flutter_meedu_videoplayer_example/pages/gridview_example.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/listview_example.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/network_with_subtitle_page.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/one_page_to_other_page_example.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/pick_file_page_example.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/playback_speed_example_page.dart';
 import 'package:flutter_meedu_videoplayer_example/pages/player_with_header_page.dart';
+import 'package:flutter_meedu_videoplayer_example/pages/portrait_example_page.dart';
 
 void main() {
   initMeeduPlayer();
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: HomePage(),
+      home: const HomePage(),
       routes: {
-        "basic": (_) => BasicExamplePage(),
-        "fullscreen": (_) => FullscreenExamplePage(),
-        "with-header": (_) => PlayerWithHeaderPage(),
-        "subtitles": (_) => NetworkWithSubtitlesPage(),
-        "playback-speed": (_) => PlayBackSpeedExamplePage(),
-        "quality-change": (_) => ChangeQualityExamplePage(),
-        "one-page-to-other": (_) => OnePageExample(),
-        "pick-file": (_) => PickFileExamplePage(),
-        "custom-icons": (_) => CustomIconsExamplePage(),
-        "disabled-buttons": (_) => DisabledButtonsExample(),
-        "listview": (_) => ListViewExample(),
+        "basic": (_) => const BasicExamplePage(),
+        "fullscreen": (_) => const FullscreenExamplePage(),
+        "with-header": (_) => const PlayerWithHeaderPage(),
+        "subtitles": (_) => const NetworkWithSubtitlesPage(),
+        "playback-speed": (_) => const PlayBackSpeedExamplePage(),
+        "quality-change": (_) => const ChangeQualityExamplePage(),
+        "one-page-to-other": (_) => const OnePageExample(),
+        "pick-file": (_) => const PickFileExamplePage(),
+        "custom-icons": (_) => const CustomIconsExamplePage(),
+        "disabled-buttons": (_) => const DisabledButtonsExample(),
+        "listview": (_) => const ListViewExample(),
+        "gridview": (_) => const GridViewExample(),
+        "portrait": (_) => const PortraitExamplePage(),
       },
     );
   }
@@ -63,67 +68,79 @@ class _HomePageState extends State<HomePage> {
             onPressed: () {
               Navigator.pushNamed(context, 'basic');
             },
-            child: Text("Basic Network example"),
+            child: const Text("Basic Network example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'fullscreen');
             },
-            child: Text("Fullscreen example"),
+            child: const Text("Fullscreen example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'with-header');
             },
-            child: Text("With header example"),
+            child: const Text("With header example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'subtitles');
             },
-            child: Text("With subtitles example"),
+            child: const Text("With subtitles example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'playback-speed');
             },
-            child: Text("Playback speed example"),
+            child: const Text("Playback speed example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'quality-change');
             },
-            child: Text("Quality Change example"),
+            child: const Text("Quality Change example"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'one-page-to-other');
             },
-            child: Text("One Page to other"),
+            child: const Text("One Page to other"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'pick-file');
             },
-            child: Text("Pick file"),
+            child: const Text("Pick file"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'custom-icons');
             },
-            child: Text("Custom Icons"),
+            child: const Text("Custom Icons"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'disabled-buttons');
             },
-            child: Text("Disabled Buttons"),
+            child: const Text("Disabled Buttons"),
           ),
           TextButton(
             onPressed: () {
               Navigator.pushNamed(context, 'listview');
             },
-            child: Text("ListView"),
+            child: const Text("ListView"),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.pushNamed(context, 'gridview');
+            },
+            child: const Text("GridView"),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.pushNamed(context, 'portrait');
+            },
+            child: const Text("Portrait"),
           )
         ],
       ),