+++ lib/src/helpers/shortcuts/short_cuts_map.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_meedu_videoplayer/src/helpers/shortcuts/actions.dart';
 import 'intents.dart';
 import 'package:flutter/services.dart';
 
@@ -13,5 +12,3 @@ Map<LogicalKeySet, Intent> shortcutsToIntents = {
   LogicalKeySet(LogicalKeyboardKey.space): const TogglePlayingStateIntent(),
   LogicalKeySet(LogicalKeyboardKey.numpadDecimal): const ToggleVideoFitIntent(),
 };
-
-