+++ lib/src/widgets/styles/primary/primary_player_controls.dart
@@ -17,6 +17,7 @@ class PrimaryVideoPlayerControls extends StatelessWidget {
     final _ = MeeduPlayerController.of(context);
 
     return ControlsContainer(
+      responsive: responsive,
       child: Stack(
         alignment: Alignment.center,
         children: [
@@ -27,13 +28,13 @@ class PrimaryVideoPlayerControls extends StatelessWidget {
               right: 0,
               top: 0,
               child: Padding(
-                padding: EdgeInsets.only(top: 10.0),
+                padding: const EdgeInsets.only(top: 10.0),
                 child: _.header!,
               ),
             ),
-          Container(
-            height: context.mediaQuerySize.height,
-            width: context.mediaQuerySize.width,
+          SizedBox(
+            height: responsive.height,
+            width: responsive.width,
           ),
 
           Row(
@@ -48,7 +49,7 @@ class PrimaryVideoPlayerControls extends StatelessWidget {
                   iconPath: 'assets/icons/rewind.png',
                   customIcon: _.customIcons.rewind,
                 ),
-                SizedBox(width: 10),
+                const SizedBox(width: 10),
               ],
               if (_.enabledButtons.playPauseAndRepeat)
                 RxBuilder(
@@ -68,7 +69,7 @@ class PrimaryVideoPlayerControls extends StatelessWidget {
                   }
                 }),
               if (_.enabledButtons.rewindAndfastForward) ...[
-                SizedBox(width: 10),
+                const SizedBox(width: 10),
                 PlayerButton(
                   onPressed: _.fastForward,
                   iconColor: Colors.white,